Move recurrence string generation to class Pretty
